Make USB3 UVC cameras work in MJPEG mode.
The default bandwith allocation is too pessimistic and won't support 4
cameras.  Hack some off it to make it work.

While we are here, add usbmon support for debugging, and setup the udev
rules to work with USB3 cameras.  They have a different path.

I still need to enable the quirk by default.

Signed-off-by: Austin Schuh <austin.linux@gmail.com>